move ToolCall to components dir
